+++ designacoes/designacoes.html
@@ -6,11 +6,11 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Designações</title>
-    <link rel="stylesheet" href="css/style.css">
+    <link rel="stylesheet" href="../modulos/style.css">
     <link rel="stylesheet" href="https://unpkg.com/intro.js/minified/introjs.min.css">
-    <script defer src="js/ui.js"></script>
+    <script defer src="../modulos/ui.js"></script>
     <script defer src="https://unpkg.com/intro.js/minified/intro.min.js"></script>
-    <script defer src="js/tour.js"></script>
+    <script defer src="../modulos/tour.js"></script>
 </head>
 <body>
 <a href="#main-content" class="skip-link">Pular para o conteúdo</a>
@@ -24,20 +24,20 @@
 <nav data-intro="Navegue pelas seções">
     <h2><span class="material-icons">settings</span> Configurações</h2>
     <ul>
-        <li><a href="index.html" title="Início"><span class="material-icons">home</span> Início</a></li>
-        <li><a href="territorios.html" title="Territórios"><span class="material-icons">description</span> Territórios</a></li>
-        <li><a href="mapa.html" title="Mapa"><span class="material-icons">map</span> Mapa</a></li>
-        <li><a href="saidas.html" title="Saídas"><span class="material-icons">directions_walk</span> Saídas</a></li>
-        <li><a href="designacoes.html" title="Designações"><span class="material-icons">push_pin</span> Designações</a></li>
-        <li><a href="calendario.html" title="Calendário"><span class="material-icons">event</span> Calendário</a></li>
-        <li><a href="sugestoes.html" title="Sugestões"><span class="material-icons">lightbulb</span> Sugestões</a></li>
-        <li><a href="relatorios.html" title="Relatórios"><span class="material-icons">bar_chart</span> Relatórios</a></li>
-        <li><a href="dashboard.html" title="Dashboard"><span class="material-icons">show_chart</span> Dashboard</a></li>
-        <li><a href="coleta.html" title="Coleta"><span class="material-icons">cloud_download</span> Coleta</a></li>
-        <li><a href="logs.html"><span class="material-icons">bug_report</span> Logs</a></li>
-        <li><a href="configuracoes.html" title="Configurações"><span class="material-icons">settings</span> Configurações</a></li>
-        <li><a href="ajuda.html" title="Ajuda"><span class="material-icons">help</span> Ajuda</a></li>
-        <li><a href="sobre.html" title="Sobre"><span class="material-icons">info</span> Sobre</a></li>
+        <li><a href="../modulos/index.html" title="Início"><span class="material-icons">home</span> Início</a></li>
+        <li><a href="../territorios/territorios.html" title="Territórios"><span class="material-icons">description</span> Territórios</a></li>
+        <li><a href="../modulos/mapa.html" title="Mapa"><span class="material-icons">map</span> Mapa</a></li>
+        <li><a href="../saidas/saidas.html" title="Saídas"><span class="material-icons">directions_walk</span> Saídas</a></li>
+        <li><a href="./designacoes.html" title="Designações"><span class="material-icons">push_pin</span> Designações</a></li>
+        <li><a href="../modulos/calendario.html" title="Calendário"><span class="material-icons">event</span> Calendário</a></li>
+        <li><a href="../modulos/sugestoes.html" title="Sugestões"><span class="material-icons">lightbulb</span> Sugestões</a></li>
+        <li><a href="../modulos/relatorios.html" title="Relatórios"><span class="material-icons">bar_chart</span> Relatórios</a></li>
+        <li><a href="../modulos/dashboard.html" title="Dashboard"><span class="material-icons">show_chart</span> Dashboard</a></li>
+        <li><a href="../modulos/coleta.html" title="Coleta"><span class="material-icons">cloud_download</span> Coleta</a></li>
+        <li><a href="../modulos/logs.html"><span class="material-icons">bug_report</span> Logs</a></li>
+        <li><a href="../modulos/configuracoes.html" title="Configurações"><span class="material-icons">settings</span> Configurações</a></li>
+        <li><a href="../modulos/ajuda.html" title="Ajuda"><span class="material-icons">help</span> Ajuda</a></li>
+        <li><a href="../modulos/sobre.html" title="Sobre"><span class="material-icons">info</span> Sobre</a></li>
     </ul>
 </nav>
 
@@ -105,8 +105,8 @@
     <span id="status-acao">📌 Aguardando ações...</span>
     <span id="status-conexao">🟢 Conectado</span>
 </footer>
-<script src="js/toast.js"></script>
-<script src="js/notificacoes.js"></script>
-<script src="js/designacoes.js"></script>
+<script src="../utils/toast.js"></script>
+<script src="../modulos/notificacoes.js"></script>
+<script src="./designacoes.js"></script>
 </body>
 </html>
