+++ saidas/Saida.js
@@ -1,7 +1,7 @@
 const { dialog } = require('electron');
 const fs = require('fs');
 const { parse } = require('csv-parse/sync');
-const db = require('./db');
+const db = require('../modulos/db');
 const logger = require('../logger');
 
 async function listar() {
