+++ territorios/territorios.html
@@ -8,9 +8,9 @@
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Gestão de Territórios - Territórios</title>
-    <link rel="stylesheet" href="css/style.css">
-    <script defer src="js/territorios.js"></script>
-    <script defer src="js/ui.js"></script>
+    <link rel="stylesheet" href="../modulos/style.css">
+    <script defer src="./territorios.js"></script>
+    <script defer src="../modulos/ui.js"></script>
 </head>
 <body>
 <header>
@@ -24,20 +24,20 @@
 <nav>
         <h2><span class="material-icons">map</span> Territórios</h2>    
     <ul>
-        <li><a href="index.html" title="Início"><span class="material-icons">home</span> Início</a></li>
-        <li><a href="territorios.html" title="Territórios"><span class="material-icons">description</span> Territórios</a></li>
-        <li><a href="mapa.html" title="Mapa"><span class="material-icons">map</span> Mapa</a></li>
-        <li><a href="saidas.html" title="Saídas"><span class="material-icons">directions_walk</span> Saídas</a></li>
-        <li><a href="designacoes.html" title="Designações"><span class="material-icons">push_pin</span> Designações</a></li>
-        <li><a href="calendario.html" title="Calendário"><span class="material-icons">event</span> Calendário</a></li>
-        <li><a href="sugestoes.html" title="Sugestões"><span class="material-icons">lightbulb</span> Sugestões</a></li>
-        <li><a href="relatorios.html" title="Relatórios"><span class="material-icons">bar_chart</span> Relatórios</a></li>
-        <li><a href="dashboard.html" title="Dashboard"><span class="material-icons">show_chart</span> Dashboard</a></li>
-        <li><a href="coleta.html" title="Coleta"><span class="material-icons">cloud_download</span> Coleta</a></li>
-        <li><a href="logs.html"><span class="material-icons">bug_report</span> Logs</a></li>
-        <li><a href="configuracoes.html" title="Configurações"><span class="material-icons">settings</span> Configurações</a></li>
-        <li><a href="ajuda.html" title="Ajuda"><span class="material-icons">help</span> Ajuda</a></li>
-        <li><a href="sobre.html" title="Sobre"><span class="material-icons">info</span> Sobre</a></li>
+        <li><a href="../modulos/index.html" title="Início"><span class="material-icons">home</span> Início</a></li>
+        <li><a href="./territorios.html" title="Territórios"><span class="material-icons">description</span> Territórios</a></li>
+        <li><a href="../modulos/mapa.html" title="Mapa"><span class="material-icons">map</span> Mapa</a></li>
+        <li><a href="../saidas/saidas.html" title="Saídas"><span class="material-icons">directions_walk</span> Saídas</a></li>
+        <li><a href="../designacoes/designacoes.html" title="Designações"><span class="material-icons">push_pin</span> Designações</a></li>
+        <li><a href="../modulos/calendario.html" title="Calendário"><span class="material-icons">event</span> Calendário</a></li>
+        <li><a href="../modulos/sugestoes.html" title="Sugestões"><span class="material-icons">lightbulb</span> Sugestões</a></li>
+        <li><a href="../modulos/relatorios.html" title="Relatórios"><span class="material-icons">bar_chart</span> Relatórios</a></li>
+        <li><a href="../modulos/dashboard.html" title="Dashboard"><span class="material-icons">show_chart</span> Dashboard</a></li>
+        <li><a href="../modulos/coleta.html" title="Coleta"><span class="material-icons">cloud_download</span> Coleta</a></li>
+        <li><a href="../modulos/logs.html"><span class="material-icons">bug_report</span> Logs</a></li>
+        <li><a href="../modulos/configuracoes.html" title="Configurações"><span class="material-icons">settings</span> Configurações</a></li>
+        <li><a href="../modulos/ajuda.html" title="Ajuda"><span class="material-icons">help</span> Ajuda</a></li>
+        <li><a href="../modulos/sobre.html" title="Sobre"><span class="material-icons">info</span> Sobre</a></li>
     </ul>
 </nav>
 
@@ -118,6 +118,8 @@
     <span id="status-acao">📌 Aguardando ações...</span>
     <span id="status-conexao">🟢 Conectado</span>
 </footer>
+
+<script src="../utils/toast.js"></script>
 
 </body>
 </html>
